# docs/openapi/jgrants-api-oas3.yaml
openapi: 3.0.0
info:
  title: 補助金情報取得API
  description: 補助金情報取得API
  version: "1.0"
paths:
  /subsidies:
    get:
      description: 補助金一覧取得API（Search subsidies with conditions）
      parameters:
        - name: keyword
          description: 検索キーワード（Keyword for
            search）<br>最小文字数は2文字とする（スペース入力不可）。<br>大文字・小文字や全角・半角の表記ゆれを許容する（例：IoTとIOT、IoTとⅠoＴ、カタカナとｶﾀｶﾅを区別しない）。
          required: true
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 255
        - name: sort
          description: ソート項目名（Field name to order
            by）<br>ソート項目名で指定した項目をソート順で並替える。<br>・created_date：作成日時<br>・acceptance_start_datetime：募集開始日時<br>・acceptance_end_datetime：募集終了日時
          required: true
          in: query
          schema:
            type: string
            enum:
              - created_date
              - acceptance_start_datetime
              - acceptance_end_datetime
        - name: order
          description: ソート順（Sort order）<br>ソート項目名で指定した項目をソート順で並替える。<br>・ASC：昇順<br>・DESC：降順
          required: true
          in: query
          schema:
            type: string
            enum:
              - ASC
              - DESC
        - name: acceptance
          description: 募集期間内絞込要否（Within an acceptance period）<br>・0：否<br>・1：要
          required: true
          in: query
          schema:
            type: string
            enum:
              - "0"
              - "1"
        - name: use_purpose
          description: 利用目的（Use of subsidies）<br>値が複数ある場合は、「 /
            」（半角スペース＋半角スラッシュ＋半角スペース）で区切る。<br>・新たな事業を行いたい<br>・販路拡大・海外展開をしたい<br>・イベント・事業運営支援がほしい<br>・事業を引き継ぎたい<br>・研究開発・実証事業を行いたい<br>・人材育成を行いたい<br>・資金繰りを改善したい<br>・設備整備・IT導入したい<br>・雇用・職場環境を改善したい<br>・エコ・SDG’s活動支援がほしい<br>・災害（自然災害、感染症等）支援がほしい<br>・教育・子育て・少子化への支援がほしい<br>・スポーツ・文化への支援がほしい<br>・安全・防災対策支援がほしい<br>・まちづくり・地域振興支援がほしい
          required: false
          in: query
          schema:
            type: string
            maxLength: 255
        - name: industry
          description: 業種（Industry）<br>値が複数ある場合は、半角スペース＋半角スラッシュ＋半角スペース）で区切る。<br>・農業，林業<br>・漁業<br>・鉱業，採石業，砂利採取業<br>・建設業<br>・製造業<br>・電気・ガス・熱供給・水道業<br>・情報通信業<br>・運輸業，郵便業<br>・卸売業，小売業<br>・金融業，保険業<br>・不動産業，物品賃貸業<br>・学術研究，専門・技術サービス業<br>・宿泊業，飲食サービス業<br>・生活関連サービス業，娯楽業<br>・教育，学習支援業<br>・医療，福祉<br>・複合サービス事業<br>・サービス業（他に分類されないもの）<br>・公務（他に分類されるものを除く）<br>・分類不能の産業
          required: false
          in: query
          schema:
            type: string
            maxLength: 255
        - name: target_number_of_employees
          description: 従業員数（Number of employees）
          required: false
          in: query
          schema:
            type: string
            enum:
              - 従業員の制約なし
              - 5名以下
              - 20名以下
              - 50名以下
              - 100名以下
              - 300名以下
              - 900名以下
              - 901名以上
        - name: target_area_search
          description: 補助対象地域（Target area to search）
          required: false
          in: query
          schema:
            type: string
            enum:
              - 全国
              - 北海道地方
              - 東北地方
              - 関東・甲信越地方
              - 東海・北陸地方
              - 近畿地方
              - 中国地方
              - 四国地方
              - 九州・沖縄地方
              - 北海道
              - 青森県
              - 岩手県
              - 宮城県
              - 秋田県
              - 山形県
              - 福島県
              - 茨城県
              - 栃木県
              - 群馬県
              - 埼玉県
              - 千葉県
              - 東京都
              - 神奈川県
              - 新潟県
              - 富山県
              - 石川県
              - 福井県
              - 山梨県
              - 長野県
              - 岐阜県
              - 静岡県
              - 愛知県
              - 三重県
              - 滋賀県
              - 京都府
              - 大阪府
              - 兵庫県
              - 奈良県
              - 和歌山県
              - 鳥取県
              - 島根県
              - 岡山県
              - 広島県
              - 山口県
              - 徳島県
              - 香川県
              - 愛媛県
              - 高知県
              - 福岡県
              - 佐賀県
              - 長崎県
              - 熊本県
              - 大分県
              - 宮崎県
              - 鹿児島県
              - 沖縄県
      responses:
        "200":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/subsidy-summaries-response"
        "400":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Bad request
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "401":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Unauthorized
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "404":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Resource not found
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "405":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Method not allowed
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "406":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Not acceptable
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "415":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Unsupported media type
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "500":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500-error-response"
        "501":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Not Implemented
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
  /subsidies/id: {}
  "/subsidies/id/{id}":
    x-amf-description: 補助金詳細取得API（Retrieve a subsidy detail）
    get:
      parameters:
        - name: id
          description: 補助金ID（Subsidy ID）
          required: true
          in: path
          schema:
            type: string
            maxLength: 18
      responses:
        "200":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/subsidy-details-response"
        "400":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Bad request
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "401":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Unauthorized
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "404":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Resource not found
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "405":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Method not allowed
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "406":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Not acceptable
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "415":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Unsupported media type
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
        "500":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500-error-response"
        "501":
          description: ""
          x-amf-mediaType: application/json
          content:
            application/json:
              schema:
                example:
                  message: Not Implemented
                type: object
                additionalProperties: true
                required:
                  - message
                properties:
                  message:
                    type: string
servers:
  - url: http://api.jgrants-portal.go.jp/exp/v1/public
  - url: https://api.jgrants-portal.go.jp/exp/v1/public
components:
  schemas:
    subsidy-summaries-response:
      description: 補助金一覧検索結果情報（Result of querying subsidies list）
      type: object
      additionalProperties: true
      required:
        - metadata
        - result
      properties:
        metadata:
          type: object
          additionalProperties: true
          required:
            - type
            - resultset
          properties:
            type:
              description: API仕様公開URL（API Specification URL）
              example: https://anypoint.mulesoft.com/exchange/497c75e7-5d6a-48b3-a39b-4105316f61f0/exp-public/
              type: string
            resultset:
              type: object
              additionalProperties: true
              required:
                - count
              properties:
                count:
                  description: 取得件数（Number of results）
                  example: 1
                  type: integer
                  format: int
        result:
          description: 補助金一覧検索結果情報（Result of querying subsidies list）
          type: array
          items:
            $ref: "#/components/schemas/type"
    500-error-response:
      description: エラー情報（Result of error）
      type: object
      additionalProperties: true
      required:
        - type
        - errorCode
        - title
        - instance
      properties:
        type:
          description: API仕様公開URL（API Specification URL）
          example: https://anypoint.mulesoft.com/exchange/497c75e7-5d6a-48b3-a39b-4105316f61f0/exp-public/
          type: string
        errorCode:
          description: エラーコード（Error code）<br>・E-ML-9999：内部サーバエラー
          example: E-ML-9999
          type: string
        title:
          description: エラータイトル（Error title）
          example: Internal Server Error
          type: string
        instance:
          description: HTTPリクエストURI（HTTP request URI）
          example: /api/subsidies
          type: string
    subsidy-details-response:
      description: 補助金詳細照会結果情報（Result of querying subsidy detail）
      type: object
      additionalProperties: true
      required:
        - metadata
        - result
      properties:
        metadata:
          type: object
          additionalProperties: true
          required:
            - type
            - resultset
          properties:
            type:
              description: API仕様公開URL（API Specification URL）
              example: https://anypoint.mulesoft.com/exchange/497c75e7-5d6a-48b3-a39b-4105316f61f0/exp-public/
              type: string
            resultset:
              type: object
              additionalProperties: true
              required:
                - count
              properties:
                count:
                  description: 取得件数（Number of results）
                  example: 1
                  type: integer
                  format: int
        result:
          description: 補助金詳細照会結果情報（Result of retrieving subsidy detail）
          type: array
          items:
            $ref: "#/components/schemas/type_1"
    type:
      description: 補助金一覧検索結果情報（Result to search subsidies list)
      type: object
      additionalProperties: true
      required:
        - id
        - name
      properties:
        id:
          description: 補助金ID（Subsidy ID）
          example: S0J0w00wer0wUgr77E
          type: string
          maxLength: 18
        name:
          description: 補助金番号（Subsidy number）
          example: S-01100011
          type: string
          maxLength: 10
        title:
          description: 補助金名（Subsidy name）
          example: 小規模事業者補助金
          x-amf-union:
            - $ref: "#/components/schemas/title"
            - type: "null"
        target_area_search:
          description: 補助対象地域（Target area to search）<br>値が複数ある場合は、「 /
            」（半角スペース＋半角スラッシュ＋半角スペース）で区切る。<br>・全国<br>・北海道地方<br>・東北地方<br>・関東・甲信越地方<br>・東海・北陸地方<br>・近畿地方<br>・中国地方<br>・四国地方<br>・九州・沖縄地方<br>・北海道<br>・青森県<br>・岩手県<br>・宮城県<br>・秋田県<br>・山形県<br>・福島県<br>・茨城県<br>・栃木県<br>・群馬県<br>・埼玉県<br>・千葉県<br>・東京都<br>・神奈川県<br>・新潟県<br>・富山県<br>・石川県<br>・福井県<br>・山梨県<br>・長野県<br>・岐阜県<br>・静岡県<br>・愛知県<br>・三重県<br>・滋賀県<br>・京都府<br>・大阪府<br>・兵庫県<br>・奈良県<br>・和歌山県<br>・鳥取県<br>・島根県<br>・岡山県<br>・広島県<br>・山口県<br>・徳島県<br>・香川県<br>・愛媛県<br>・高知県<br>・福岡県<br>・佐賀県<br>・長崎県<br>・熊本県<br>・大分県<br>・宮崎県<br>・鹿児島県<br>・沖縄県
          example: 東京都 / 大阪府
          x-amf-union:
            - $ref: "#/components/schemas/target_area_search_response"
            - type: "null"
        subsidy_max_limit:
          description: 補助額上限（Subsidy maximum amount）<br>入力値以下の補助額上限が設定されている補助金に絞り込む。
          example: 10000000
          x-amf-union:
            - $ref: "#/components/schemas/subsidy_max_limit"
            - type: "null"
        acceptance_start_datetime:
          description: 募集開始日時（Acceptance start date）
          example: 2020-02-28T16:41:41.090Z
          x-amf-union:
            - $ref: "#/components/schemas/acceptance_start_datetime"
            - type: "null"
        acceptance_end_datetime:
          description: 募集終了日時（Acceptance end date）
          example: 2021-02-28T16:41:41.090Z
          x-amf-union:
            - $ref: "#/components/schemas/acceptance_end_datetime"
            - type: "null"
        target_number_of_employees:
          description: 従業員数（Number of employees）
          example: 20名以下
          x-amf-union:
            - $ref: "#/components/schemas/target_number_of_employees"
            - type: "null"
    type_1:
      description: 補助金詳細照会結果情報（Result of querying subsidy detail）
      type: object
      additionalProperties: true
      required:
        - id
        - name
      properties:
        id:
          description: 補助金ID（Subsidy ID）
          example: S0J0w00wer0wUgr77E
          type: string
          maxLength: 18
        name:
          description: 補助金番号（Subsidy number）
          example: S-01100011
          type: string
          maxLength: 10
        title:
          description: 補助金名（Subsidy name）
          example: 小規模事業者補助金
          x-amf-union:
            - $ref: "#/components/schemas/title"
            - type: "null"
        subsidy_catch_phrase:
          description: 補助金のキャッチコピー（Advertising slogan）
          example: 小規模事業者の生産性向上と持続的発展を図る
          x-amf-union:
            - $ref: "#/components/schemas/subsidy_catch_phrase"
            - type: "null"
        detail:
          description: 補助金のサマリー（Purpose / Overview）
          example: 小規模事業者が取り組む販路開拓等の取組の経費の一部を補助することにより、生産性向上と持続的発展を図ることを目的とします。
          x-amf-union:
            - $ref: "#/components/schemas/detail"
            - type: "null"
        use_purpose:
          description: 利用目的（Use of subsidies）<br>値が複数ある場合は、「 /
            」（半角スペース＋半角スラッシュ＋半角スペース）で区切る。<br>・新たな事業を行いたい<br>・販路拡大・海外展開をしたい<br>・イベント・事業運営支援がほしい<br>・事業を引き継ぎたい<br>・研究開発・実証事業を行いたい<br>・人材育成を行いたい<br>・資金繰りを改善したい<br>・設備整備・IT導入したい<br>・雇用・職場環境を改善したい<br>・エコ・SDG’s活動支援がほしい<br>・災害（自然災害、感染症等）支援がほしい<br>・教育・子育て・少子化への支援がほしい<br>・スポーツ・文化への支援がほしい<br>・安全・防災対策支援がほしい<br>・まちづくり・地域振興支援がほしい
          example: 新たな事業を行いたい / 設備整備・IT導入したい
          x-amf-union:
            - $ref: "#/components/schemas/use_purpose"
            - type: "null"
        industry:
          description: 業種（Industry）<br>値が複数ある場合は、半角スペース＋半角スラッシュ＋半角スペース）で区切る。<br>・農業，林業<br>・漁業<br>・鉱業，採石業，砂利採取業<br>・建設業<br>・製造業<br>・電気・ガス・熱供給・水道業<br>・情報通信業<br>・運輸業，郵便業<br>・卸売業，小売業<br>・金融業，保険業<br>・不動産業，物品賃貸業<br>・学術研究，専門・技術サービス業<br>・宿泊業，飲食サービス業<br>・生活関連サービス業，娯楽業<br>・教育，学習支援業<br>・医療，福祉<br>・複合サービス事業<br>・サービス業（他に分類されないもの）<br>・公務（他に分類されるものを除く）<br>・分類不能の産業
          example: 情報通信業 / 教育，学習支援業
          x-amf-union:
            - $ref: "#/components/schemas/industry"
            - type: "null"
        target_area_search:
          description: 補助対象地域（Target area to search）<br>値が複数ある場合は、「 /
            」（半角スペース＋半角スラッシュ＋半角スペース）で区切る。<br>・全国<br>・北海道地方<br>・東北地方<br>・関東・甲信越地方<br>・東海・北陸地方<br>・近畿地方<br>・中国地方<br>・四国地方<br>・九州・沖縄地方<br>・北海道<br>・青森県<br>・岩手県<br>・宮城県<br>・秋田県<br>・山形県<br>・福島県<br>・茨城県<br>・栃木県<br>・群馬県<br>・埼玉県<br>・千葉県<br>・東京都<br>・神奈川県<br>・新潟県<br>・富山県<br>・石川県<br>・福井県<br>・山梨県<br>・長野県<br>・岐阜県<br>・静岡県<br>・愛知県<br>・三重県<br>・滋賀県<br>・京都府<br>・大阪府<br>・兵庫県<br>・奈良県<br>・和歌山県<br>・鳥取県<br>・島根県<br>・岡山県<br>・広島県<br>・山口県<br>・徳島県<br>・香川県<br>・愛媛県<br>・高知県<br>・福岡県<br>・佐賀県<br>・長崎県<br>・熊本県<br>・大分県<br>・宮崎県<br>・鹿児島県<br>・沖縄県
          example: 東京都 / 大阪府
          x-amf-union:
            - $ref: "#/components/schemas/target_area_search_response"
            - type: "null"
        target_area_detail:
          description: 補助対象地域詳細（Target area detail）
          example: 全国
          x-amf-union:
            - $ref: "#/components/schemas/target_area_detail"
            - type: "null"
        target_number_of_employees:
          description: 従業員数（Number of employees）
          example: 20名以下
          x-amf-union:
            - $ref: "#/components/schemas/target_number_of_employees"
            - type: "null"
        subsidy_rate:
          description: 補助率（Subsidy rate）
          example: 20%
          x-amf-union:
            - $ref: "#/components/schemas/subsidy_rate"
            - type: "null"
        subsidy_max_limit:
          description: 補助額上限（Subsidy maximum amount）<br>入力値以下の補助額上限が設定されている補助金に絞り込む。
          example: 10000000
          x-amf-union:
            - $ref: "#/components/schemas/subsidy_max_limit"
            - type: "null"
        acceptance_start_datetime:
          description: 募集開始日時（Acceptance start date）
          example: 2020-02-28T16:41:41.090Z
          x-amf-union:
            - $ref: "#/components/schemas/acceptance_start_datetime"
            - type: "null"
        acceptance_end_datetime:
          description: 募集終了日時（Acceptance end date）
          example: 2021-02-28T16:41:41.090Z
          x-amf-union:
            - $ref: "#/components/schemas/acceptance_end_datetime"
            - type: "null"
        project_end_deadline:
          description: 事業終了期限（Project end deadline）
          example: 2020-07-31T15:00:00.000Z
          x-amf-union:
            - $ref: "#/components/schemas/project_end_deadline"
            - type: "null"
        request_reception_presence:
          description: 申請受付有無（Request acceptability）<br>・有：申請を受付ける<br>・無：申請を受付けない
          example: 有
          x-amf-union:
            - $ref: "#/components/schemas/request_reception_presence"
            - type: "null"
        is_enable_multiple_request:
          description: 複数回申請可否（Multiple requests acceptability）<br>・true：申請可<br>・false：申請不可
          example: false
          x-amf-union:
            - $ref: "#/components/schemas/is_enable_multiple_request"
            - type: "null"
        front_subsidy_detail_page_url:
          description: 事業者向け補助金詳細画面URL
          example: https://jgrants-2-xxx/subsidy/999999999999999999
          x-amf-union:
            - $ref: "#/components/schemas/front_subsidy_detail_page_url"
            - type: "null"
        application_guidelines:
          description: 公募要領（Application guidelines）
          example:
            - name: 公募要領.pdf
              data: JVBERi0xLjQKJe
          x-amf-union:
            - $ref: "#/components/schemas/application_guidelines"
            - type: "null"
        outline_of_grant:
          description: 交付要綱（Outline of grant）
          example:
            - name: 交付要綱.pdf
              data: JVBERi0xLjQKJe
          x-amf-union:
            - $ref: "#/components/schemas/outline_of_grant"
            - type: "null"
        application_form:
          description: 申請様式（Application form）
          example:
            - name: 申請様式.pdf
              data: JVBERi0xLjQKJe
          x-amf-union:
            - $ref: "#/components/schemas/application_form"
            - type: "null"
    title:
      type: string
      maxLength: 100
    target_area_search_response:
      type: string
      maxLength: 1000
    subsidy_max_limit:
      type: number
      maximum: 999999999999999900
    acceptance_start_datetime:
      type: string
      format: date-time
    acceptance_end_datetime:
      type: string
      format: date-time
    target_number_of_employees:
      enum:
        - 従業員の制約なし
        - 5名以下
        - 20名以下
        - 50名以下
        - 100名以下
        - 300名以下
        - 900名以下
        - 901名以上
      type: string
    subsidy_catch_phrase:
      type: string
      maxLength: 255
    detail:
      type: string
      maxLength: 131072
    use_purpose:
      type: string
      maxLength: 255
    industry:
      type: string
      maxLength: 255
    target_area_detail:
      type: string
      maxLength: 255
    subsidy_rate:
      type: string
      maxLength: 255
    project_end_deadline:
      type: string
      format: date-time
    request_reception_presence:
      enum:
        - 有
        - 無
      type: string
    is_enable_multiple_request:
      type: boolean
    front_subsidy_detail_page_url:
      type: string
      maxLength: 100
    application_guidelines:
      type: array
      items:
        $ref: "#/components/schemas/file_data"
    outline_of_grant:
      type: array
      items:
        $ref: "#/components/schemas/file_data"
    application_form:
      type: array
      items:
        $ref: "#/components/schemas/file_data"
    file_data:
      type: object
      additionalProperties: true
      properties:
        name:
          description: ファイル名（File name）
          example: ガイドライン.pdf
          type: string
          maxLength: 255
        data:
          description: データ（File data）<br>Base64形式
          type: string